+++ test/support/helpers/businesses_helper.rb
@@ -1,0 +1,12 @@
+module BusinessesHelper
+  def business_attributes
+    {
+      user: users(:empty),
+      name: "Name",
+      company: "Company",
+      bio: "Bio",
+      avatar: active_storage_blobs(:basecamp),
+      developer_notifications: :no
+    }
+  end
+end
